[survey] modify automatically redirect to workflow-choice page logic

+++ tailwindui/app/(dashboard)/dashboard/page.tsx
@@ -130,7 +130,7 @@ export default function Dashboard() {
       console.log('projects', projects);
 			setProjects(projects);
 			setRendered(true);
-			if (projects.length == 0) {
+			if (projects.length == 0 && !showSurvey) {
 				sessionStorage.clear();
 				router.push('/workflow-type-choice');
 			}
